+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Build navigation easily, without any efforts. Library that provides tools to bui
 # 0.0.2
 
 1. Rename `Item` class to `LinkContainer`
+2. Visitor for menu part. Allows to run functionality outside the menu part, e.g. to make modifications.
 
 # 0.0.1